+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All changes to the **Hearts of Modding** extension will be documented in this fi
 
 - **Slight refactor to reduce scanner boilerplate code.**
 
+- **Better `common/countries/*.txt` scanning.**
+
 ## [v0.21.0] - 2026-06-15
 
 > its becoming a bit too burdensome